improved responsive layouts. cleanup code.
