+++ next.config.js
@@ -9,6 +9,8 @@ const NEXT_PUBLIC_SERVER_URL = process.env.NEXT_PUBLIC_SERVER_URL
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   images: {
+    deviceSizes: [640, 768, 1024, 1280, 1536, 1920, 2048, 3840],
+    imageSizes: [32, 48, 64, 96, 128, 256, 384],
     qualities: [75, 100],
     remotePatterns: [
       ...[NEXT_PUBLIC_SERVER_URL].map((item) => {
@@ -56,6 +58,7 @@ const nextConfig = {
   },
   serverExternalPackages: ['pdf-img-convert', 'pdfjs-dist', 'canvas'],
   transpilePackages: ['react-pdf', 'flipbook-js'],
+  allowedDevOrigins: ['192.168.50.7'],
 }
 
 export default withPayload(nextConfig, { devBundleServerPackages: false })
