+++ src/Header/Component.client.tsx
@@ -5,10 +5,13 @@ import { ColorThemeToggle } from '@/providers/Theme/color-theme-toggle'
 import type { Header } from '@/payload-types'
 import { useHeaderTheme } from '@/providers/HeaderTheme'
 import { usePathname } from 'next/navigation'
-import NavMenu from '@/components/NavMenu'
 import { NavTreeItem } from '@/utilities/buildNavTree'
 import { cn } from '@/lib/utils'
 import { AppMainLogo } from '@/components/AppMainLogo'
+import { useIsMobile } from '@/hooks/use-mobile'
+import MobileNavMenu from '@/components/NavMenu/MobileNavMenu'
+import HeaderNavMenu from '@/components/NavMenu/HeaderNavMenu'
+import { PageTOCTriggerButton } from '@/components/PageTableOfContents'
 
 interface HeaderClientProps {
   data: Header
@@ -17,7 +20,7 @@ interface HeaderClientProps {
   twitchStatusSlot?: React.ReactNode
 }
 
-export const HeaderRowStyles = 'px-2 py-2 mx-auto grid grid-cols-12 grid-rows-1 gap-3 rounded-none'
+export const HeaderRowStyles = 'grid grid-cols-12 grid-rows-1 gap-3 rounded-none px-3 '
 
 export const HeaderClient: React.FC<HeaderClientProps> = ({
   appTitle,
@@ -30,6 +33,7 @@ export const HeaderClient: React.FC<HeaderClientProps> = ({
   const { headerThemeMode, setHeaderThemeMode, headerThemeColor, setHeaderThemeColor } =
     useHeaderTheme()
   const pathname = usePathname()
+  const isMobile = useIsMobile()
 
   useEffect(() => {
     setHeaderThemeMode(null)
@@ -47,31 +51,51 @@ export const HeaderClient: React.FC<HeaderClientProps> = ({
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [headerThemeColor])
 
-  return (
-    <header
-      className={`fixed top-0 z-20 w-screen rounded-none bg-background shadow-md`}
-      data-theme={themeColor}
-      data-mode={themeMode}
-    >
-      {/* TOP ROW OF HEADER */}
-      <div className={cn(HeaderRowStyles, 'border-b')}>
-        <div className="col-span-6 flex justify-start align-middle md:col-span-4">
-          {twitchStatusSlot}
+  if (isMobile) {
+    return (
+      <header
+        className={`fixed top-0 z-20 w-screen rounded-none bg-background shadow-md`}
+        data-theme={themeColor}
+        data-mode={themeMode}
+      >
+        <div className={cn(HeaderRowStyles, 'border-b py-2')}>
+          <div className="col-span-2 flex flex-nowrap items-center justify-start">
+            <MobileNavMenu appTitle={appTitle} navTree={navTree} />
+          </div>
+          <div className="col-span-8 flex flex-nowrap items-center justify-center">
+            <AppMainLogo text={appTitle} variant={'default'} />
+          </div>
+          <div className="col-span-2 flex flex-nowrap items-center justify-end">
+            <PageTOCTriggerButton size={'lg'} />
+          </div>
         </div>
-        <div className="col-span-6 flex flex-row flex-nowrap justify-center rounded-none md:col-span-4">
-          <AppMainLogo text={appTitle} />
+      </header>
+    )
+  } else {
+    return (
+      <header
+        className={`fixed top-0 z-20 w-screen rounded-none bg-background shadow-md`}
+        data-theme={themeColor}
+        data-mode={themeMode}
+      >
+        {/* TOP ROW OF HEADER */}
+        <div className={cn(HeaderRowStyles, 'border-b py-1')}>
+          <div className="col-span-4 flex items-center justify-start">{twitchStatusSlot}</div>
+          <div className="col-span-4 flex flex-nowrap items-center justify-center rounded-none">
+            <AppMainLogo text={appTitle} />
+          </div>
+          <div className="col-span-4 flex items-center justify-end">
+            <ColorThemeToggle />
+          </div>
         </div>
-        <div className="hidden justify-end md:col-span-4 md:flex">
-          <ColorThemeToggle />
-        </div>
-      </div>
 
-      {/* BOTTOM ROW OF HEADER */}
-      <div className={cn(HeaderRowStyles, 'container')}>
-        <div className="col-span-full flex flex-row flex-nowrap justify-center rounded-none bg-background/95 backdrop-blur supports-backdrop-filter:bg-background/60 xl:col-span-10 xl:col-start-3">
-          <NavMenu navTree={navTree} appTitle={appTitle} />
+        {/* BOTTOM ROW OF HEADER */}
+        <div className={cn(HeaderRowStyles, 'container py-1')}>
+          <div className="col-span-full flex flex-row flex-nowrap justify-center rounded-none bg-background/95 backdrop-blur supports-backdrop-filter:bg-background/60 xl:col-span-10 xl:col-start-2">
+            <HeaderNavMenu navTree={navTree} />
+          </div>
         </div>
-      </div>
-    </header>
-  )
+      </header>
+    )
+  }
 }
